feat: allow uploading and exporting at the same time (#26)
Now one can both export data to a local dir and upload to a remote
server. This allows taking the output and uploading to flamegraph.com
using https://github.com/pyroscope-io/flamegraph.com-github-action

This change is a slightly breaking one, but I believe no one should be
relying on the previous behaviour (where `--noUpload` also makes it
write to a local dir).